% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appshiny.R
\name{riks_api}
\alias{riks_api}
\title{One of those shiny}
\usage{
riks_api(rate1 = "", rate2 = "", rate3 = "", rate4 = "",
  rate5 = "")
}
\arguments{
\item{input}{One parameter}

\item{output}{Another parameter+-}
}
\description{
One of those shiny
}
