% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appshiny_api_rc.R
\docType{class}
\name{shiny_api-class}
\alias{shiny_api-class}
\alias{shiny_api}
\title{One of those shiny}
\arguments{
\item{rate1}{Us maturity 3 month}

\item{rate2}{Japanese maturity 3 month}

\item{rate3}{Greate britan maturity 3 month}

\item{rate4}{Eurocountries maturity 3 month}

\item{rate5}{Norwegian maturity 3 month}
}
\description{
One of those shiny
}
\section{Fields}{

\describe{
\item{\code{us_rate}}{character.}

\item{\code{eu_rate}}{character.}

\item{\code{nor_rate}}{character.}

\item{\code{uk_rate}}{character.}

\item{\code{jp_rate}}{character.}

\item{\code{base}}{character.}

\item{\code{from}}{character.}

\item{\code{to}}{character.}

\item{\code{adres}}{character.}

\item{\code{repo_day}}{data.frame.}

\item{\code{test}}{data.frame.}

\item{\code{start}}{numeric.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(start = "Welcome")}}{Starts the api and initilize the values.}

\item{\code{riks_api(us = "a", eu = "a", nor = "a", uk = "a", jp = "")}}{Method Centralbank API conects to central bank of sweden to get the data.}

\item{\code{shiny_app()}}{Shiny_app starts the shiny and illustrates the data.}
}}

