% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appshiny.R
\name{riks_api}
\alias{riks_api}
\title{One of those shiny}
\usage{
riks_api(rate1 = "", rate2 = "", rate3 = "", rate4 = "",
  rate5 = "")
}
\arguments{
\item{rate1}{Us maturity 3 month}

\item{rate2}{Japanese maturity 3 month}

\item{rate3}{Greate britan maturity 3 month}

\item{rate4}{Eurocountries maturity 3 month}

\item{rate5}{Norwegian maturity 3 month}
}
\description{
One of those shiny
}
